Updated the spec, added Makefile to build html rfc
Added Makefile to build html and txt versions of the rfc.

Added the node xml tag to the spec.  Modified the spec to have link use
the node tag.

Added the Error message to the spec.  Added a list of standard Error codes.

Added the Status message to the spec.  Added the neighbors element for status.

Added the concept of connection subtypes.  Defined structured.near and
structured.shortcut

imported from: boykin@pobox.com--brunetdev-2005/brunet--pobbranch--0.1--patch-1