Added IXmlAble interface
Added an interface called IXmlAble.  It is interface for the Xml support
that our objects provide for serialization.  Made ConnectionMessage implement
this interface (which meant adding 2 methods).  Added code for all
the ConnectionMessage objects (LinkMessage, PingMessage...) to support this.
The reason for this is mostly to make Xml support uniform across classes
and not tied to being a ConnectionMessage.  It also simplifies writing
unit tests.

imported from: boykin@pobox.com--brunetdev-2005/brunet--pobbranch--0.1--patch-4