Added NUnit tests for all ConnectionMessage classes
Added NUnit tests for all ConnectionMessage classes and the
ConnectionMessageParser.

Removed the CpTester class from tests directory since its functionality
is now in the test fixtures in the classes themselves.

To build a testable build, use "nant test".  There are preprocessor flags
that will prevent the test code from being built without it.  This is
to ensure that NUnit is not required to build the code, but only to run
the tests.  Finally, nunit-console Brunet.dll will run all the tests
and print a report.

imported from: boykin@pobox.com--brunetdev-2005/brunet--pobbranch--0.1--patch-5