Double lock fix
This *should* work to fix the double lock issues.  It is a little hard to test
this patch, but it seems to work.

I tested by using BootStrapTester with 2 nodes.  I had it print when the
double lock condition happened.  It seemed to happen and retry, so I guess
that is good.

It remains to be seen if this will affect PlanetLab networks which may
or may not be having any double lock problems.

imported from: boykin@pobox.com--brunetdev-2005/brunet--pobbranch--0.1--patch-11