Fixed some exception cases in TcpEdgeListener and Linker
I added some improved Exception handling in TcpEdgeListener and Linker.

In TcpEdgeListener there were two issues I saw:
1) TryNextIP would sometimes throw an exception.  This was due to the
Remove method being called with a null key.  This could happen when
the method could not create a socket, and so the socket was still null.
I added a check if( s != null ) before we try to remove it.

2) In Linker, when we could not get more TransportAddress objects, an
exception was printed.  This exception does not need to be printed,
so the print statement was wrapped with a #if debug type thing.

3) In the SelectLoop in TcpEdgeListener, Accept can fail with an exception.
This exception was uncaught before and would result in the TcpEdgeListener
stopping its listening thread.  This is very bad.  We now catch that
exception, however, we ignore it (which means an incoming connection is
just lost).  We should probably consider under what circumstances
TcpEdgeListener should quit (indeed how can EdgeListeners stop listening
due to errors?)

imported from: boykin@pobox.com--brunetdev-2005/brunet--pobbranch--0.1--patch-14