Added disconnection support to BootStrapTester
When running BootStrapTester, you can enter the index of a node to disconnect.
This should test to see if the system removes the node and if the other nodes
correctly deal with that nodes departure.

When you are done, type "Q" to quit, and all nodes will be disconnected.

Made Disconnect GracefullyClose all the connections.

Made TcpEdgeListener Close Sockets it is done with

imported from: boykin@pobox.com--brunetdev-2005/brunet--pobbranch--0.1--patch-15