Continued cleaning ConnectionTable, added StatusMessage to Connection
Continued to clean ConnectionTable.  Changed the ConnectionEventArgs constructor
which was only called in ConnectionTable.  Now the ConnectionEventArgs carries
the Connection, as well as the old interface.  The advantage is by accessing
the Connection you can get access to the StatusMessage that was passed to
the Node.  This should allow connecting directly to its neighbors.

imported from: boykin@pobox.com--brunetdev-2005/brunet--pobbranch--0.1--patch-17