Added GetConnections method to ConnectionTable
Add the GetConnections method to ConnectionTable.  This allows one to do
a foreach(Connection c in tab.GetConnections("structured")) { }.  This
is the prefered way of accessing the ConnectionTable when you want
all edges or connections of a given type.

Added SimpleConnectionOverlord, a proposed rewrite of SCO.  Needs more
testing (especially with trimming, which is now disabled).

imported from: boykin@pobox.com--brunetdev-2005/brunet--pobbranch--0.1--patch-20