Settled on a definition for right and left neighbors, Random adds BST
A right neighbor is someone that is closer to the right than left
and is one of the k closest nodes to the right (since the ring wraps
around this second condition is needed).  There is some small probability
(high with non-random addresses) that this definition will break down.
Thus it is important to have many nodes with random addresses to make
sure this assertion is likely to be true.

Made BootStrapTester use randomly generated addresses.  Had to modify
the output code to make sure each node is aliased with a small integer
in the correct order.

Updated graph_check.py to work with the new BootStrapTester output.

imported from: boykin@pobox.com--brunetdev-2005/brunet--pobbranch--0.1--patch-21