Fixed some bugs in Linker and Connector
This is a bugfix patch.  It does (basically) three things:

1) Fixes several bugs related to the double-lock fix that previously
was added.  It had some problems, and this fixes some of them.

2) A small (possible) bug existed in Connector regarding thread
synchronization.  Fixed by this patch.  We still need to audit
the code class by class and look for problems of this sort.

3) I changed the status message to include all nodes of the same
Maintype as the connection.  This will cause problems if there are
more neighbors than can fit in a packet so we need a better patch
that will only send the nearest nodes to the other side (or something...)

imported from: boykin@pobox.com--brunetdev-2005/brunet--pobbranch--0.1--patch-22