Changed all the XML parsing to SAX rather than DOM
Used stream based parsing for XML since we do it very often.  Unfortunately,
it is not as much faster as I had hoped (only 20-40% faster not several times
faster).  Maybe a better SAX parser in a later version of Mono will speed
things up.  Right now it takes about 70ms on my 1.8 GHz K7 to parse a
message.  Since it takes about 200-400 ms to handle an AHPacket, this is
a significant fraction of that time.

imported from: boykin@pobox.com--brunetdev-2005/brunet--pobbranch--0.1--patch-27