PacketCallback delegate was replaced with IPacketHandler interface
Changed the delegates in the packet handling loop to Interfaces.

Some benchmarks show delegates to be as much as 4 times slower than
interfaces, however for us it was not a clear win.  The biggest
win for us is that using delegates makes the profile information from
the mono runtime (--profile) difficult to trace (since there are
some internal calls associated with delegates that wind up
being the most important elements in the profile).

imported from: boykin@pobox.com--brunetdev-2005/brunet--pobbranch--0.1--patch-30