Timeout bug fixes, Linker increases timeout
There was a bug in Node that made it never close an edge due to a timeout.

In fact, several variables we were using were never initialized, so even
after fixing the bug in Node, it exposed other bugs in UdpEdge and TcpEdge.

Seems to work as expected now, edges which have been inactive for 4 minutes
are closed.  Connections with no activity are pinged after 30 seconds.

I made linker increase its timeout after each timeout.  So, it slows down
when there are failures.  This seems to improve the situation a lot
when the system is loaded, but I can't tell for sure.   A plab test
will be valuable.

These fixes seem to work for me.  I don't know if they will help the
highly loaded cases.

imported from: boykin@pobox.com--brunetdev-2005/brunet--pobbranch--0.1--patch-36