Better TransportAddress handling, harvesting.
Connections now include the LinkMessage their peer sent them to
start the connection.  This information allows the Node to
add the TransportAddress our peer thinks we are to our list
of LocalTAs.

Added an Properties to Edge so that we can check to see
if a Local or RemoteTA is ephemeral.  If it is not ephemeral
we can harvest this TransportAddress.

Node now watches all new Connections and harvests local and
remote TransportAddresses where appropriate.  This should
make UDP behind NAT support work much better (as well as STUN),

Handled the case where a node has a misconfigured hostname.
Everything should work fine as long as it is configured with
correct RemoteTAs to start with (it can learn its IP address
from incoming connections).

imported from: boykin@pobox.com--brunetdev-2005/brunet--pobbranch--0.1--patch-43