UdpEdge fixes and Linker Timeout changes
* I upped the _restart_attempts in Linker to 16 from 4 (this
is a totally random number).  We need to do some calculations,
or some experiments to fix a good number for this, or come up
with an adaptive algorithm to set it properly.  I also
reduced the restart probability to 10%.  This broadens the
exponential distribution on how long it takes to restart.  This
should reduce the probability that a double-lock situation is
followed by a second double lock.

* I changed the way UdpEdges work slightly.  The new way should
prevent a weird condition of an edge closing on one side and
being reopened as a new edge for one side, but still the old
edge for the other.

imported from: boykin@pobox.com--brunetdev-2005/brunet--pobbranch--0.1--patch-56