Mainline Merge, Fixed a minor bug in RwtaRouter and Cleaned UCO
Merged with mainline

RwtaRouter was not delivering locally when there were no leafs.
This meant when the network was being bootstrapped, no one got
the packets sent to unstructured addresses.  The fix is a one
liner.

Also I added some thread synchronization to UCO and while I was
at it I simplified some of the code.

imported from: boykin@pobox.com--brunetdev-2005/brunet--pobbranch--0.1--patch-58