Got rid of NearestMulti, things appear to work.
Got rid of the NearestMulti concept.  There was no way to differentiate
the Nearest behavior from NearestMulti.  Specifically, there was no
way to do it without changing the routing algorithm based on the options,
and that is going too far.

The Last mode is what you want when you want to send a packet to one
structured address that is nearest (or second nearest) a particular
value.  This is how we make shortcuts.

imported from: boykin@pobox.com--brunetdev-2005/brunet--pobexp--0.1--patch-5