SimpleConnectionOverlord makes use of CtM neighbor lists
The key here is when the Connector finishes, the SCO checks to see
if any of the neighbors of the responders should be neighbors of
the requester.  If they should, SCO tries to connect to the closest
left and right neighbor.

This should guarantee that firewalled nodes get properly connected since
each node learns about pairs of nodes which are connected.

imported from: boykin@pobox.com--brunetdev-2005/brunet--pobexp--0.1--patch-8