Added ADR Serialization docs and code, and RPC system based on it.
Added ADR Serialization.  This serializes simple types (like strings, numbers,
booleans, lists, hashtables, exceptions...) to a simple (guessable) binary
format.  This is used for RPC in the RpcManager.

Added RpcManager to allow for RPC calls over Brunet.  There is an example
and test code in the rpc_test.boo file.

Added BlockingQueue.  This is a valuable thread synchonization tool used in
RpcManager.  It is basically a thread-safe queue that can block on dequeue
to be woken when an object is enqueued.  It provides an "object pipe" from one
thread to another.

imported from: boykin@pobox.com--code-2006/brunet--mainline--0.3--patch-1