Rewrite of Ipop - No longer do we have this concept of significantly different nodes, they all are inherited from IpopNode - Connectivity to Brunet is no longer part of the IpopNodes job, we use BasicNode for this purpose (could be renamed BrunetNode) - Abstracted away all the typical mechanisms used by an IpopNode such as packet parsing, ethernet reading, and communication with Brunet - User has the choice to implement their own Multicast, DNS, and DHCP handlers, thus making a ShaIpopNode should be relatively easy (i.e., the original IPRouter) - Removed the notion of having all these different Dht / Xmlrpc Server notions, this should be ported to Brunet and reside there, currently it is now known as DhtService.dll embedding IBlockingQueue, the Client, and Server features. - BasicNode contains all the libraries previously in IPOP_Common and Brunet_Common, this should be ported to Brunet - Currently ctrl-c for Linux is not implemented
I am open to suggestions on the coding style, I spent a long time trying to tweak this into some OO format.  I originally wanted to read from the Ethernet and just pass to the thread in Brunet, but due to the fact that ICopyable does not allow accessing without copying first, decided not to (I just had a revelation that my previous thoughts on the matter may be wrong).  This is still a possibility though.

The main purpose of this rewrite was to make adding encryption handlers and routing based upon rpc injection easier.  In otherwords, implementing DNAT / SNAT (packet transformation with routing table) of iptables and security in IPsec.

This design also should make things easier to test.