Tunnel uses SyncMessage instead of SyncMessages and "AHSender" instead of forwarding sender
Since SyncMessages only focus on connections that are near to the
requesting address and they are only active for nearest neighbors they
aren't too useful for most connections.  For this, I have implemented a
sync message that can be defined in the ITunnelOverlap.  Users define
both the SyncMessage and the processor for the StatusMessage.  At this
point, logic for a Full Tunnel Overlap or a NCService Overlap Is
trivial.

The problem with using forwarding senders is that they tend to bloat up
a perfectly fine sized packet as it traverses the system.  It is better
to simply forward the packet ourselves via an edge an AHHeader (as done
in the previous tunnel) than attempt to use a ForwardingSender.  Note,
we still use a forwarding sender in the outbound edge until it hears
back from the remote in bound edge so that the inbound edge knows of a
guaranteed route back to the outbound edge.

Implementing a new ITunnelOverlap.FindOverlap will allow nodes to force a
ITunnelOverlap.EvaluatePotentialOverlap if there are no good (for whatever
reason) Overlap.